+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-04-11
+### Added
+- Backing up Group Policy Configurations (Administrative Templates) added.
+- Backing up Group Policy Configuration (Administrative Template) Assignments added.
+- Restoring Group Policy Configurations (Administrative Templates) added.
+- Restoring Group Policy Configuration (Administrative Template) Assignments added.
+
 ## [1.1.0] - 2019-03-17
 ### Added
 - Added function `Compare-IntuneBackupFile`